backend/job_trackers/pyproject.toml
[project]
name = "job_trackers"
version = "0.1.0"
description = "job-trackers using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.118.0,<1.0.0",
    "tavily-python>=0.7.3",
]

[project.scripts]
job_trackers = "job_trackers.main:run"
run_crew = "job_trackers.main:run"
train = "job_trackers.main:train"
replay = "job_trackers.main:replay"
test = "job_trackers.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
